# Performance: Optimize Batch Operations with Parallel Processing